create listing

+++ src/main/java/org/tiere/dto/Listing.java
@@ -4,5 +4,5 @@ import org.tiere.util.ListingType;
 
 import java.util.List;
 
-public record Listing(Long id, List<Animal> animals, ListingType type, List<Long> files) {
+public record Listing(Integer id, List<Animal> animals, ListingType type, List<Integer> files) {
 }
